Improve odd-sized magic square generation, update README
Improve how odd-sized magic squares are generated.

Update the README for a better explanation on the parameters used
to generate odd-sized matrix squares.

Fix a bug where not enough memory would be allocated for a matrix.

Fix a bug where the memory would not be freed at the end of the
program.